TS: extracting NumberedPaginationLink from Table/Footer
